+++ src/test/java/com/stylemycloset/notification/event/listener/FeedCommentNotificationEventListenerIntegrationTest.java
@@ -18,7 +18,7 @@ import com.stylemycloset.ootd.repo.FeedRepository;
 import com.stylemycloset.sse.dto.SseInfo;
 import com.stylemycloset.sse.repository.SseRepository;
 import com.stylemycloset.sse.service.impl.SseServiceImpl;
-import com.stylemycloset.testutil.IntegrationTestSupport;
+import com.stylemycloset.IntegrationTestSupport;
 import com.stylemycloset.user.entity.User;
 import com.stylemycloset.user.repository.UserRepository;
 import java.util.Deque;
